Update readme to reflect modern work
My main user readme has been referring to C++ for the past few years, but I've been working on Rust and Go for over a year. This updates the project to reflect the current state of things.

Will update this later to include links to my most recent Rust projects as well.